Add product DTOs mapper and tests

+++ app/src/main/java/com/thejasvee/coolblue/domain/model/ProductReviewSummary.kt
@@ -2,5 +2,5 @@ package com.thejasvee.coolblue.domain.model
 
 data class ProductReviewSummary(
     val average: Double,
-    val count: Double,
+    val count: Int,
 )
